major refactoring
I tried to separate the View and the Model. I don't want to commit to
a GUI architecture, yet, as I'm not sure what DataBinding is and how
useful that is to my use case.

The Wisdom of the Internet tries to push me towards MVVM, but I think
that is not needed here.